+++ HISTORY.md
@@ -2,6 +2,8 @@
 
 - Re-add checks for pre-bgzipped fastq inputs to alignment preparation thanks to
   a fix for grabix to handle Illumina bgzip outputs.
+- Provide DNA damage filtering for low frequency sequencing errors in somatic
+  samples. Use `tools_on: [damage_filter]`
 - Improve lumpy structural variant runs from pre-aligned BAM files, using
   extract_sv_reads to avoid need to resort input files. Thanks to Neill Gibson.
 - Speed up post-alignment callability calculations by using default parameters